feat(bitcoin): introduce hash-like types
Types like "Txid", "Wtxid", "BlockHash", "DescriptorId", "TxMerkleNode" are all just 32 byte
arrays that represent hashes with different meanings. Currently they
are represented as strings at the FFI layer, but they are also
meaningful arrays of bytes. Particularly if a user wants to implement
persistence over the FFI layer, they would want to efficiently
serialize these types.

Here I am introducing a new group of types that all implement display,
allow serialization to bytes, and may be constructed from an array of
bytes. I went with a "rule of 3s" here, and also introduced a macro to
do these implementations because there was a lot of boilerplate
involved.

Note that all of these are included in the wallet changeset, which is
required to represent in-full for FFI-layer custom persistence.