refactor!: convert `ChangeSet` to an object type
The BDK `ChangeSet` offers some nice methods for building and working
with changesets, particularly the `merge` method. Unfortunately to
implement `merge` on `ChangeSet`, as well as the other constructors, we
must convert the `ChangeSet` type to an object instead of a record. It
is best to do this now instead of having to break the API in the future.