feat(kyoto): update to latest version
A couple new `Info` variants are added in this release to give the user
more context as to what is happening with the block header chain. There
is also a new function, `lookup_host`, that allows the developer to
query Bitcoin DNS seeders. Because the user may already configure a DNS
resolver, this resolver is added to the `CbfClient` and used to resolve
DNS queries.

There was also missing docs for adding a Socks5 proxy.